requirements.txt
azure-identity>=1.16.0
openai>=1.50.2
openreview-py>=1.26.2
pymupdf>=1.24.10
Pillow>=10.4.0
